requirements.txt
celery==4.0.2
cloudinary==1.8.0
coverage==4.4.1
coveralls==1.1
dj-database-url==0.4.2
Django==1.11.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-mailgun==0.9.1
djangorestframework==3.6.3
docopt==0.6.2
ipython==6.0.0
psycopg2==2.7.1
python-env==1.0.0
sendgrid==4.1.0
